separated spec files for video popup
