+++ cypress/integration/v2/blocks/video-popup/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Video Popup Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
