+++ cypress/integration/v2/blocks/video-popup/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Video Popup Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
