+++ cypress/integration/v2/blocks/video-popup/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Video Popup Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
